pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skywalker"
version = "0.27.0"
description = "GCP Audit & Reporting Tool for UCR Research Computing ('Ursa Major' Compliance)"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Charles Forsyth", email = "charles.forsyth@ucr.edu" }
]
license = { text = "MIT" }

# The "Skywalker" Stack
dependencies = [
    "google-cloud-compute>=1.0.0",
    "google-cloud-resource-manager>=1.0.0",
    "google-cloud-asset>=3.0.0",
    "google-cloud-storage>=2.0.0",
    "google-cloud-filestore>=1.0.0",
    "google-cloud-monitoring>=2.0.0",
    "google-cloud-container>=2.0.0",
    "google-cloud-aiplatform>=1.0.0",
    "google-cloud-notebooks>=1.0.0",
    "google-cloud-run>=0.10.0",
    "google-cloud-functions>=1.0.0",
    "google-cloud-iam>=2.0.0",
    "google-api-python-client>=2.0.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "tenacity>=8.0.0",
    "rich>=13.0.0",
    "weasyprint>=60.0.0",
    "jinja2>=3.1.0",
    "humanize>=4.0.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0"
]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pandas-stubs>=2.0.0"
]

[project.scripts]
# This creates the 'skywalker' command in your shell after install
skywalker = "skywalker.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/skywalker"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "RET", "SIM", "ARG", "PTH", "RUF"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["skywalker"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true


